Rename Entities -> Resources::Publishable
